+++ jparty/constants.py
@@ -1,1 +1,12 @@
-DEBUG = False
+from PyQt6.QtGui import QColor, QPalette
+
+FJTIME = 31
+QUESTIONTIME = 4
+MONIES = [[200, 400, 600, 800, 1000], [400, 800, 1200, 1600, 2000]]
+
+
+JBLUE = QColor("#1010a1")
+DARKBLUE = QColor("#0b0b74")
+CARDPAL = QPalette()
+CARDPAL.setColor(QPalette.ColorRole.Window, JBLUE)
+CARDPAL.setColor(QPalette.ColorRole.WindowText, QColor("#ffffff"))
