+++ jparty/environ.py
@@ -1,5 +1,4 @@
 import sys
-import os
 
 if getattr(sys, "frozen", False):
     root = getattr(sys, "_MEIPASS", ".")  # os.path.dirname(sys.executable)
