+++ jparty/style.py
@@ -1,53 +1,16 @@
-import sys
-import os
-from PyQt6.QtGui import (
-    QImage,
-    QPainter,
-    QPen,
-    QBrush,
-    QColor,
-    QFont,
-    QMovie,
-    QPixmap,
-    QDesktopServices,
-    QPalette,
-    QGuiApplication,
-    QFontDatabase,
-    QColor,
-)
+from PyQt6.QtWidgets import QStyle, QCommonStyle
+from PyQt6.QtGui import QPalette, QColor
+from PyQt6.QtCore import Qt
 
-import requests
-
-
-# from PyQt6.QtMultimedia import QSound
-from PyQt6.QtWidgets import *  # QWidget, QApplication, QDesktopWidget, QPushButton
-from PyQt6.QtCore import Qt, QRectF, QRect, QPoint, QTimer, QSize, QDir, QMargins, pyqtSignal
-import logging
-
-import pickle
-from threading import Thread, active_count
-import time
-import subprocess
-
-import threading
-from functools import partial
-
-# from .data_rc import *
-from .retrieve import get_game, get_game_sum
-from .controller import BuzzerController
-from .boardwindow import DisplayWindow, HostDisplayWindow
-from .game import Player, Game
-from .constants import DEBUG
-from .utils import SongPlayer, resource_path, check_internet
-from .logger import qt_exception_hook
+from jparty.utils import DynamicLabel, add_shadow
 
 
 class JPartyStyle(QCommonStyle):
     PM_dict = {
-        QStyle.PixelMetric.PM_LayoutBottomMargin : 0,
-        QStyle.PixelMetric.PM_LayoutLeftMargin : 0,
-        QStyle.PixelMetric.PM_LayoutRightMargin : 0,
-        QStyle.PixelMetric.PM_LayoutTopMargin : 0,
+        QStyle.PixelMetric.PM_LayoutBottomMargin: 0,
+        QStyle.PixelMetric.PM_LayoutLeftMargin: 0,
+        QStyle.PixelMetric.PM_LayoutRightMargin: 0,
+        QStyle.PixelMetric.PM_LayoutTopMargin: 0,
         QStyle.PixelMetric.PM_LayoutHorizontalSpacing: 0,
         QStyle.PixelMetric.PM_LayoutVerticalSpacing: 0,
     }
@@ -56,9 +19,23 @@ class JPartyStyle(QCommonStyle):
     }
 
     def pixelMetric(self, key, *args, **kwargs):
-        return JPartyStyle.PM_dict.get(key,
-                                       super().pixelMetric(key, *args, **kwargs))
+        return JPartyStyle.PM_dict.get(key, super().pixelMetric(key, *args, **kwargs))
 
     def styleHint(self, key, *args, **kwargs):
-        return JPartyStyle.SH_dict.get(key,
-                                       super().styleHint(key, *args, **kwargs))
+        return JPartyStyle.SH_dict.get(key, super().styleHint(key, *args, **kwargs))
+
+
+class MyLabel(DynamicLabel):
+    def __init__(self, text, initialSize, parent=None):
+        super().__init__(text, initialSize, parent)
+        self.font().setBold(True)
+        self.setWordWrap(True)
+        self.setAlignment(Qt.AlignmentFlag.AlignCenter)
+
+        add_shadow(self)
+
+        palette = QPalette()
+        palette.setColor(QPalette.ColorRole.WindowText, QColor("white"))
+        self.setPalette(palette)
+
+        self.show()
